Updated blockFunctions and merkle
Merkle:
Deleted inline script element that i was using for testing
Put quotations around javascript call to onclick on the hash it button and added the value that I needed to make it work.

BlockFunctions:
Added an alertText variable to display in the case where a user does not input anything in the texbox. Prior behavior was that the paragraph would display the text "undefined."