Updated index.html and merkle.html.
index.html: deleted the top id as the go to top button seems to work without it

merkle.html: added some sections and some anchor links. The anchor links work, but the sticky nav bar hides the content of the titles of the section.
Current work-around is to put the anchor before the title.